+++ src/index.ts
@@ -1,36 +1,17 @@
 import './tasks/compile';
 import './tasks/prepend_spdx_license';
-import { extendConfig } from 'hardhat/config';
-import 'hardhat/types/config';
+import './type-extensions.js';
 
-declare module 'hardhat/types/config' {
-  interface HardhatUserConfig {
-    spdxLicenseIdentifier?: {
-      overwrite?: boolean;
-      runOnCompile?: boolean;
-      only?: string[];
-      except?: string[];
-    };
-  }
+// import { extendConfig } from 'hardhat/config';
 
-  interface HardhatConfig {
-    spdxLicenseIdentifier: {
-      overwrite: boolean;
-      runOnCompile: boolean;
-      only: string[];
-      except: string[];
-    };
-  }
-}
-
-extendConfig((config, userConfig) => {
-  config.spdxLicenseIdentifier = Object.assign(
-    {
-      overwrite: false,
-      runOnCompile: false,
-      only: [],
-      except: [],
-    },
-    userConfig.spdxLicenseIdentifier,
-  );
-});
+// extendConfig((config, userConfig) => {
+//   config.spdxLicenseIdentifier = Object.assign(
+//     {
+//       overwrite: false,
+//       runOnCompile: false,
+//       only: [],
+//       except: [],
+//     },
+//     userConfig.spdxLicenseIdentifier,
+//   );
+// });
