+++ src/type-extensions.ts
@@ -1,0 +1,14 @@
+import {
+  SpdxLicenseIdentifierConfig,
+  SpdxLicenseIdentifierUserConfig,
+} from './types.js';
+
+declare module 'hardhat/types/config' {
+  interface HardhatConfig {
+    spdxLicenseIdentifier: SpdxLicenseIdentifierConfig;
+  }
+
+  interface HardhatUserConfig {
+    spdxLicenseIdentifier?: SpdxLicenseIdentifierUserConfig;
+  }
+}
