+++ src/types.ts
@@ -1,0 +1,9 @@
+export type SpdxLicenseIdentifierConfig = {
+  overwrite: boolean;
+  runOnCompile: boolean;
+  only: string[];
+  except: string[];
+};
+
+export type SpdxLicenseIdentifierUserConfig =
+  Partial<SpdxLicenseIdentifierConfig>;
